[ds-heap] broken down into functions, header, main
